refactor: harden product-safe image boundaries
