+++ docs/backend-contracts.md
@@ -252,6 +252,7 @@ app/api/images/cleanup-previous/route.ts
 ```
 
 For a complete App Router recipe, see [Next.js draft lifecycle](./recipes/nextjs-draft-lifecycle.md).
+For the submit transaction pattern, see [Product submit with image draft](./recipes/product-submit-with-image-draft.md).
 
 ## Package defaults stay opt-in
